Refactor http source configuration to a separate http source common package.
