+++ data-prepper-plugins/http-source-common/build.gradle
@@ -8,17 +8,11 @@ plugins {
 }
 
 dependencies {
-    implementation project(':data-prepper-api')
-    implementation project(':data-prepper-plugins:blocking-buffer')
-    implementation project(':data-prepper-plugins:http-source-common')
     implementation project(':data-prepper-plugins:common')
-    implementation project(':data-prepper-plugins:armeria-common')
     implementation libs.armeria.core
-    implementation libs.commons.io
     implementation 'software.amazon.awssdk:acm'
     implementation 'software.amazon.awssdk:s3'
     implementation 'software.amazon.awssdk:apache-client'
-    testImplementation project(':data-prepper-api').sourceSets.test.output
     testImplementation 'org.assertj:assertj-core:3.25.3'
 }
