+++ data-prepper-plugins/http-source-common/src/main/java/org/opensearch/dataprepper/http/codec/Codec.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.source.loghttp.codec;
+package org.opensearch.dataprepper.http.codec;
 
 import com.linecorp.armeria.common.HttpData;
