+++ data-prepper-plugins/http-source-common/src/test/java/org/opensearch/dataprepper/http/LogThrottlingRejectHandlerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.source.loghttp;
+package org.opensearch.dataprepper.http;
 
 import org.opensearch.dataprepper.metrics.PluginMetrics;
 import com.linecorp.armeria.common.AggregatedHttpResponse;
