+++ data-prepper-plugins/http-source-common/src/test/java/org/opensearch/dataprepper/http/certificate/CertificateProviderFactoryTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.source.loghttp.certificate;
+package org.opensearch.dataprepper.http.certificate;
 
 import org.hamcrest.core.IsInstanceOf;
 import org.junit.jupiter.api.BeforeEach;
@@ -12,7 +12,7 @@ import org.opensearch.dataprepper.plugins.certificate.CertificateProvider;
 import org.opensearch.dataprepper.plugins.certificate.acm.ACMCertificateProvider;
 import org.opensearch.dataprepper.plugins.certificate.file.FileCertificateProvider;
 import org.opensearch.dataprepper.plugins.certificate.s3.S3CertificateProvider;
-import org.opensearch.dataprepper.plugins.source.loghttp.HTTPSourceConfig;
+import org.opensearch.dataprepper.http.HttpServerConfig;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
@@ -22,21 +22,21 @@ class CertificateProviderFactoryTest {
     private final String TEST_SSL_CERTIFICATE_FILE = getClass().getClassLoader().getResource("test_cert.crt").getFile();
     private final String TEST_SSL_KEY_FILE = getClass().getClassLoader().getResource("test_decrypted_key.key").getFile();
 
-    private HTTPSourceConfig httpSourceConfig;
+    private HttpServerConfig httpServerConfig;
     private CertificateProviderFactory certificateProviderFactory;
 
     @BeforeEach
     void setUp() {
-        httpSourceConfig = mock(HTTPSourceConfig.class);
+        httpServerConfig = mock(HttpServerConfig.class);
     }
 
     @Test
     void getCertificateProviderFileCertificateProviderSuccess() {
-        when(httpSourceConfig.isSsl()).thenReturn(true);
-        when(httpSourceConfig.getSslCertificateFile()).thenReturn(TEST_SSL_CERTIFICATE_FILE);
-        when(httpSourceConfig.getSslKeyFile()).thenReturn(TEST_SSL_KEY_FILE);
+        when(httpServerConfig.isSsl()).thenReturn(true);
+        when(httpServerConfig.getSslCertificateFile()).thenReturn(TEST_SSL_CERTIFICATE_FILE);
+        when(httpServerConfig.getSslKeyFile()).thenReturn(TEST_SSL_KEY_FILE);
 
-        certificateProviderFactory = new CertificateProviderFactory(httpSourceConfig);
+        certificateProviderFactory = new CertificateProviderFactory(httpServerConfig);
         final CertificateProvider certificateProvider = certificateProviderFactory.getCertificateProvider();
 
         assertThat(certificateProvider, IsInstanceOf.instanceOf(FileCertificateProvider.class));
@@ -44,12 +44,12 @@ class CertificateProviderFactoryTest {
 
     @Test
     void getCertificateProviderS3ProviderSuccess() {
-        when(httpSourceConfig.isSslCertAndKeyFileInS3()).thenReturn(true);
-        when(httpSourceConfig.getAwsRegion()).thenReturn("us-east-1");
-        when(httpSourceConfig.getSslCertificateFile()).thenReturn("s3://data/certificate/test_cert.crt");
-        when(httpSourceConfig.getSslKeyFile()).thenReturn("s3://data/certificate/test_decrypted_key.key");
+        when(httpServerConfig.isSslCertAndKeyFileInS3()).thenReturn(true);
+        when(httpServerConfig.getAwsRegion()).thenReturn("us-east-1");
+        when(httpServerConfig.getSslCertificateFile()).thenReturn("s3://data/certificate/test_cert.crt");
+        when(httpServerConfig.getSslKeyFile()).thenReturn("s3://data/certificate/test_decrypted_key.key");
 
-        certificateProviderFactory = new CertificateProviderFactory(httpSourceConfig);
+        certificateProviderFactory = new CertificateProviderFactory(httpServerConfig);
         final CertificateProvider certificateProvider = certificateProviderFactory.getCertificateProvider();
 
         assertThat(certificateProvider, IsInstanceOf.instanceOf(S3CertificateProvider.class));
@@ -57,11 +57,11 @@ class CertificateProviderFactoryTest {
 
     @Test
     void getCertificateProviderAcmProviderSuccess() {
-        when(httpSourceConfig.isUseAcmCertificateForSsl()).thenReturn(true);
-        when(httpSourceConfig.getAwsRegion()).thenReturn("us-east-1");
-        when(httpSourceConfig.getAcmCertificateArn()).thenReturn("arn:aws:acm:us-east-1:account:certificate/1234-567-856456");
+        when(httpServerConfig.isUseAcmCertificateForSsl()).thenReturn(true);
+        when(httpServerConfig.getAwsRegion()).thenReturn("us-east-1");
+        when(httpServerConfig.getAcmCertificateArn()).thenReturn("arn:aws:acm:us-east-1:account:certificate/1234-567-856456");
 
-        certificateProviderFactory = new CertificateProviderFactory(httpSourceConfig);
+        certificateProviderFactory = new CertificateProviderFactory(httpServerConfig);
         final CertificateProvider certificateProvider = certificateProviderFactory.getCertificateProvider();
 
         assertThat(certificateProvider, IsInstanceOf.instanceOf(ACMCertificateProvider.class));
