+++ data-prepper-plugins/http-source-common/src/test/java/org/opensearch/dataprepper/http/codec/JsonCodecTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.source.loghttp.codec;
+package org.opensearch.dataprepper.http.codec;
 
 import com.linecorp.armeria.common.HttpData;
 import org.junit.jupiter.api.Test;
