+++ settings.gradle
@@ -173,3 +173,4 @@ include 'data-prepper-plugins:http-common'
 include 'data-prepper-plugins:flatten-processor'
 include 'data-prepper-plugins:mongodb'
 include 'data-prepper-plugins:rds-source'
+include 'data-prepper-plugins:http-source-common'
